fix: fix lib TCA6408A type err and build script
- Note that the build script has issue that can't export
  -DLIB_USE_XXX_PLATFORM to global_env. Will be fixed in following
  commits.

+++ lib/devices/TCA6408A/platform/dummy/tca6408a_dummy.c
@@ -24,7 +24,7 @@
 #include "tca6408a_dummy.h"
 
 tca6408a_err_t tca6408a_reg_info_to_str(const tca6408a_reg_info_t* info,
-                                        char* str, size_t max_len)
+                                        char* str, uint32_t max_len)
 {
     return 0;
 }
